requirements.txt
apache-airflow
kaggle 
torch
torchvision
pillow
opencv-python-headless
scikit-learn
mlflow
flask